refactor(cli): hoist helper ssh/provider/image/workspaceinfo to prefixed sub-packages
