+++ cmd/internal/helperimage/get_image_platforms.go
@@ -1,4 +1,4 @@
-package image
+package helperimage
 
 import (
 	"context"
