+++ cmd/internal/helperprovider/check_provider_update.go
@@ -1,4 +1,4 @@
-package provider
+package helperprovider
 
 import (
 	"bytes"
