+++ cmd/internal/helperssh/ssh_client.go
@@ -1,4 +1,4 @@
-package ssh
+package helperssh
 
 import (
 	"context"
