+++ cmd/internal/helperssh/ssh_git_clone.go
@@ -1,4 +1,4 @@
-package ssh
+package helperssh
 
 import (
 	"context"
