+++ cmd/internal/helperssh/ssh_server.go
@@ -1,4 +1,4 @@
-package ssh
+package helperssh
 
 import (
 	"encoding/base64"
