+++ cmd/internal/helperworkspaceinfo/get_workspace_config.go
@@ -1,4 +1,4 @@
-package workspaceinfo
+package helperworkspaceinfo
 
 import (
 	"context"
