+++ cmd/internal/helperworkspaceinfo/get_workspace_name.go
@@ -1,4 +1,4 @@
-package workspaceinfo
+package helperworkspaceinfo
 
 import (
 	"context"
