+++ cmd/internal/helperworkspaceinfo/get_workspace_uid.go
@@ -1,4 +1,4 @@
-package workspaceinfo
+package helperworkspaceinfo
 
 import (
 	"context"
